Display player IDs
We now display each player's ID in their box, along with the text
"(me!)" identifying the client's player. I've also taken the liberty of
reducing the grid size in order to improve readability. If the
dimensions are changed again, we will have to manually update some of
the styles for the players -- in the future, we should compute the
player styles based on the grid dimensions to avoid this (perhaps using
a library like "styled-components").