Prep for deployment
I needed to make a few adjustments to the Dockerfile(s) in order to
install native node modules in the alpine container. I also moved the
`redux-devtools-extension` package from the development dependencies to
the (real-deal) dependencies. Obviously this is not ideal, but it is a
quick fix for now, and it does not impact performance. In the future, I
will add a .env file and only load the devtools in development mode.